// types.ts
export enum Theme {
  VOID = 'VOID',
  TECH = 'TECH',
  MEDICAL = 'MEDICAL',
  FINANCIAL = 'FINANCIAL',
  MAINSTREAM = 'MAINSTREAM',
  SCIENTIFIC = 'SCIENTIFIC',
  TABLOID = 'TABLOID',
  // New Themes
  RADIOLOGY = 'RADIOLOGY',
  BIOTECH = 'BIOTECH',
  PHARMA = 'PHARMA',
  TERMINAL = 'TERMINAL',
  LIFESTYLE = 'LIFESTYLE',
  VINTAGE = 'VINTAGE',
  ACADEMIC = 'ACADEMIC',
  GLOBAL_NETWORK = 'GLOBAL_NETWORK',
  LEGAL = 'LEGAL',
  TECH_BLOG = 'TECH_BLOG',
  GOV_REPORT = 'GOV_REPORT',
  NEURO = 'NEURO',
  CARDIOLOGY = 'CARDIOLOGY',
  ONCOLOGY = 'ONCOLOGY',
  // Latest Additions
  CYBER = 'CYBER',
  SPACE = 'SPACE',
  CLIMATE = 'CLIMATE',
  CRYPTO = 'CRYPTO',
  QUANTUM = 'QUANTUM',
}